+++ sandbox/loader/js/lib4.js
@@ -1,4 +1,0 @@
-YUI.add('lib4',function(Y) {
-  
-},'@VERSION@', {  })
-