+++ src/handlebars/js/yui-handlebars-utils.js
@@ -30,8 +30,9 @@ Handlebars.SafeString.prototype.toString = function () {
 
 Handlebars.Utils = {
     escapeExpression: function (string) {
-        if (!string) {
-            return '';
+        // Skip escaping for empty strings.
+        if (string === '') {
+            return string;
         }
 
         // Don't escape SafeStrings, since they're already (presumed to be)