+++ src/yui/tests/core-tests.js
@@ -38,6 +38,13 @@ YUI.add('core-tests', function(Y) {
 
         name: "Core tests",
 
+        _should: {
+            ignore: {
+                'getLocation() should return the location object': Y.UA.nodejs,
+                'getLocation() should return `undefined` when executing in node.js': !Y.UA.nodejs
+            }
+        },
+
         'cached functions should execute only once per input': function() {
 
             var r1 = "", r2 = "", r3 = "";
@@ -96,8 +103,16 @@ YUI.add('core-tests', function(Y) {
 
         },
 
+        'getLocation() should return the location object': function () {
+            Y.Assert.areSame(Y.config.win.location, Y.getLocation(), 'Did not return Y.config.win.location.');
+        },
+
+        'getLocation() should return `undefined` when executing in node.js': function () {
+            Y.Assert.isUndefined(Y.getLocation(), 'Did not return `undefined`');
+        },
+
         test_cached_undefined_null: function() {
-            
+
             var f1 = function(a) {
                 return a;
             };
@@ -109,7 +124,7 @@ YUI.add('core-tests', function(Y) {
             Y.Assert.areEqual(a, undefined);
             a = c1('foo');
             Y.Assert.areEqual(a, 'foo');
-            
+
         },
 
         test_ie_enum_bug: function() {
@@ -124,7 +139,7 @@ YUI.add('core-tests', function(Y) {
         },
         test_guid: function() {
             var id, id2, i;
-            for (i = 0; i < 1000; i++) {           
+            for (i = 0; i < 1000; i++) {
                 id = Y.guid();
                 id2 = Y.guid();
                 Y.Assert.areNotEqual(id, id2, 'GUID creation failed, ids match');
@@ -132,7 +147,7 @@ YUI.add('core-tests', function(Y) {
         },
         test_stamp: function() {
             var id, id2, i;
-            for (i = 0; i < 1000; i++) {           
+            for (i = 0; i < 1000; i++) {
                 id = Y.stamp({});
                 id2 = Y.stamp({});
                 Y.Assert.areNotEqual(id, id2, 'STAMP GUID creation failed, ids match');
@@ -152,6 +167,19 @@ YUI.add('core-tests', function(Y) {
                 Assert.isObject(Y.Node, 'Node was not loaded');
             });
         },
+        /* TODO Find a better way to test this.
+        This test will fail since it's loading all modules.
+        Including test modules so the test will re-execute and
+        loop de loop..
+
+        test_use_star: function() {
+            var Assert = Y.Assert,
+                testY = YUI().use('*');
+
+            Assert.isObject(testY.Test, 'Failed to load via use *');
+            
+        },
+        */
         test_one_submodule: function() {
             var Assert = Y.Assert;
             YUI({
@@ -203,7 +231,7 @@ YUI.add('core-tests', function(Y) {
                   logMe: true,
                   butNotMe: false
               };
-              
+
               Y.log('test logInclude logMe','info','logMe');
               Assert.areEqual(last, 'logMe', 'logInclude (true) Failed');
               last = undefined;
@@ -223,24 +251,6 @@ YUI.add('core-tests', function(Y) {
 
             });
             console.info = l;
-        },
-        test_loader_combo_sep: function() {
-            var Assert = Y.Assert;
-            
-            var testY = YUI({
-                debug: true,
-                combine: true,
-                comboSep: '--;;--'
-            });
-            var loader = new testY.Loader(testY.config);
-            testY.Env._loader = loader;
-
-            testY.Get.script = function(s) {
-                var url = s[0];
-                Assert.isTrue((url.indexOf(testY.config.comboSep) > 5), 'Combo seperator (' + testY.config.comboSep + ') was not used');
-                testY._loading = false;
-            };
-            testY.use('editor');
         },
         test_global_apply_config: function() {
             var Assert = Y.Assert,
@@ -266,14 +276,56 @@ YUI.add('core-tests', function(Y) {
             Assert.isObject(YUI.GlobalConfig.modules, 'modules object in global config not created');
             Assert.isObject(YUI.GlobalConfig.modules.davglass, 'First module in global config not created');
             Assert.isObject(YUI.GlobalConfig.modules.foo, 'Second module in global config not created');
+
+        },
+        test_old_config: function() {
+            var Assert = Y.Assert,
+                testY = YUI();
+
+            testY._config({
+                foo: true
+            });
+
+            Assert.isTrue(testY.config.foo, 'Old style config did not apply');
+        },
+        test_applyTo: function() {
+            var Assert = Y.Assert,
+                testY = YUI(),
+                id = testY.id;
+
+            testY.io = {
+                xdrReady: function() {
+                    Assert.areSame(id, this.id, 'Did not apply to the right instance');
+                }
+            };
+
+            testY.foobar = function() {
+                Assert.isTrue(false, 'testY.foobar should not have ever fired.');
+            };
+
+            YUI.applyTo(id, 'io.xdrReady', []); //Should call
+            YUI.applyTo(id, 'io.xdrResponse', []); //Does not exist
+            YUI.applyTo(id, 'foobar', []); //Should not call
+            YUI.applyTo('1234567890', 'io.xdrReady', []); //Should not call since instance id is invalid
+        },
+        test_global_config: function() {
+            var Assert = Y.Assert,
+                test = this;
             
+            YUI().use('global-mod', function(Y) {
+                test.resume(function() {
+                    Assert.isTrue(Y.GlobalMod, 'Module in global config failed to load');
+                });
+            });
+
+            test.wait(3000);
         },
         test_multiple_ua: function() {
             var Assert = Y.Assert,
                 globalUA = YUI.Env.UA,
                 uaFF = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10.7; rv:6.0) Gecko/20100101 Firefox/6.0',
                 uaChrome = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_7_0) AppleWebKit/535.2 (KHTML, like Gecko) Chrome/15.0.860.0 Safari/535.2';
-    
+
             var Y1 = YUI();
             Assert.areEqual(globalUA.userAgent, Y1.UA.userAgent, 'Global UA and local (Y1) UA are different');
 
@@ -287,6 +339,91 @@ YUI.add('core-tests', function(Y) {
 
             Assert.areEqual(YUI.Env.UA.userAgent, Y2.UA.userAgent, 'Global UA and local (Y2) UA are different');
 
+        },
+        test_conditional: function() {
+            var Assert = Y.Assert,
+            test = this;
+
+            YUI.add('cond', function(Y) {
+                Y.cond = true;
+            });
+
+            YUI.add('cond-test', function(Y) {
+                Y.condTest = true;
+            });
+
+            YUI({
+                modules: {
+                    'cond-test': {
+                        condition: {
+                            trigger: 'cond',
+                            test: function() {
+                                return true;
+                            }
+                        }                       
+                    }
+                }
+            }).use('cond', function(Y2) {
+                //Should be sync..
+                Assert.isTrue(Y2.cond, 'Conditional module was not loaded.');
+                Assert.isTrue(Y2.condTest, 'Conditional module was not loaded.');
+            });
+            
+        },
+        test_missed: function() {
+            var Assert = Y.Assert;
+
+            var testY = YUI().use('bogus');
+
+            Assert.areSame(1, testY.Env._missed.length, 'Failed to error on bogus module');
+            Assert.areSame('bogus', testY.Env._missed[0], 'Failed to error on bogus module');
+
+            YUI.add('bogus', function(Y) { Y.bogus = true; });
+
+            testY.use('bogus');
+
+            Assert.areSame(0, testY.Env._missed.length, 'Failed to remove added bogus module');
+
+        },
+        test_attach_error: function() {
+            var Assert = Y.Assert;
+            YUI.add('attach-error', function() { Y.push(); });
+
+            YUI({
+                errorFn: function(str) {
+                    Assert.isTrue(str.indexOf('attach-error') > -1, 'Failed to fire errorFn on attach error');
+                    return true;
+                }
+            }).use('attach-error');
+        },
+        test_attach_after: function() {
+            var Assert = Y.Assert;
+            YUI.add('after-test', function(Y) {
+                Y.afterTest = true; 
+                Assert.isObject(Y.Node, 'Node not loaded before this module');
+            }, '1.0.0', {
+                after: [ 'node' ]
+            });
+
+            YUI().use('after-test', function(Y2) {
+                Assert.isTrue(Y2.afterTest, 'after-test module was not loaded');
+            });
+        },
+        test_dump_core: function() {
+            var Assert = Y.Assert,
+                o = {},
+                testY = YUI();
+
+            var str = testY.dump(o);
+            Assert.isString(str, 'Default Y.dump failed to return a string');
+        },
+        test_destroy: function() {
+            var Assert = Y.Assert,
+                testY = YUI();
+            
+            testY.destroy();
+            Assert.isUndefined(testY.Env, 'Environment not destroyed');
+            Assert.isUndefined(testY.config, 'Instance config not destroyed');
         }
     });